refactor: Code cleanup due C4 and other audit feedback (SC-2039) (#370)
* chore: C4 events updates

- Comment indicates that FundsWithdrawn event should be emitted only when _withdrawableDividend > 0  (https://github.com/code-423n4/2021-04-maple-findings/issues/96)
- No event is emitted (https://github.com/code-423n4/2021-04-maple-findings/issues/8)
- Missing event for critical operation of setAdmin change for Protocol admin in MapleGlobals.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/38)
- Missing event for critical operation of Pool Delegate validity change in MapleGlobals.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/39)
- Missing event for critical operation of setAdmin change in Pool.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/69)
- Missing index on events (https://github.com/code-423n4/2021-04-maple-findings/issues/86)
- Function triggerDefault should call _emitBalanceUpdateEventForCollateralLocker (https://github.com/code-423n4/2021-04-maple-findings/issues/99)
- Missing event for critical operation of new Collateral locker creation in CollateralLockerFactory.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/27)
- Missing event for critical operation of new Debt locker creation in DebtLockerFactory.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/29)
- Missing event for critical operation of new Funding locker creation in FundingLockerFactory.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/30)
- Missing event for critical operation of new Liquidity locker creation in LiquidityLockerFactory.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/31)
- Missing event for critical operation of setAdmin change in Loan.sol (https://github.com/code-423n4/2021-04-maple-findings/issues/33)

* chore: Normalize event indexing and comments

- Normalized event declaration spacing
- Added @dev comment for all events thrown in functions
- Cleaned up @dev comments in BasicFDT and ExtendedFDT
- Added indexing for all reasonably filterable event parameters

* chore: explicit and differentiated admin names

Mirrored admin variables in global context, Pool, PoolFactory, Loan and LoanFactory may make it confusing for deployment and maintenance (https://github.com/code-423n4/2021-04-maple-findings/issues/36)

- Loan.admins -> Loan.loanAdmins
- Loan.setAdmin -> Loan.setLoanAdmin
- LoanFactory.admins -> LoanFactory.loanFactoryAdmins
- LoanFactory.setAdmin -> LoanFactory.setLoanFactoryAdmin
- MapleGlobals.admin -> MapleGlobals.globalAdmin
- MapleGlobals.setAdmin -> MapleGlobals.setGlobalAdmin
- Pool.admin -> Pool.poolAdmins
- Pool.setAdmin -> Pool.setPoolAdmin
- PoolFactory.admin -> PoolFactory.poolFactoryAdmins
- PoolFactory.setAdmin -> PoolFactory.setPoolFactoryAdmin

* refactor: cleanup and correct interfaces

Interface and implementation function declaration differs (https://github.com/code-423n4/2021-04-maple-findings/issues/98)

- Corrected non-matching interface names
- Explicit uint256 usage
- Removed unused or non-existent interfaces
- Increase interface coverage
- Remove unused variables
- Set minimal explicit state-mutabilities
- Standardized "visibility | state-mutability" pattern for functions
- Fixed one different compiler version

* refactor: FDT interface inheritance

- Fixed circular naming of FDT and IFDT
- Created Interfaces for all FDT variants (in their own directory)
- FDT interface inheritance all the way up to shed redundant interface declarations

* chore: BPTVal and getPoolSharesRequired warning

Functions calculating the value of BPT is vulnerable to flash-loan attacks (https://github.com/code-423n4/2021-04-maple-findings/issues/113)

* chore: PR pre-review fixes

- remove unwanted event arg spacing

* refactor: Normalize admin set NatSpec and arguments

* revert: PR review fixes

- Fix Alignment
- Save gas optimizations for separate PR

* chore: PR Review for alignment

+++ contracts/interfaces/IBPool.sol
@@ -57,8 +57,8 @@ interface IBPool {
 
     function exitswapExternAmountOut(
         address tokenOut,
-        uint tokenAmountOut,
-        uint maxPoolAmountIn
-    ) external returns (uint poolAmountIn);
+        uint256 tokenAmountOut,
+        uint256 maxPoolAmountIn
+    ) external returns (uint256 poolAmountIn);
 
 }
